Refactor several things into Layer
